requirements.txt
flask-cors
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
decorator==5.1.1
executing==1.2.0
Flask==2.2.2
flask-restx==1.0.5
gunicorn==20.1.0
ipython
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.0
stack-data==0.6.2
toml==0.10.2
traitlets==5.8.1
typing_extensions==4.4.0
wcwidth==0.2.6
Werkzeug==2.2.2
tqdm
mysql-connector-python
python-dotenv
boto3
botocore
requests
pandas
docxtpl
openpyxl
redis